Reinstate govuk_delivery in the backend class
We are trying to remove govuk_delivery as it has been retired. The
resources have been marked `ensure => 'absent'` within the
govuk_delivery app but because it was also removed from the backend
class that file was not being executed.

This commit reinstates the app so as puppet can remove it.