Remove govuk_delivery mongo env
As part of my ongoing struggle to get rid of govuk_delivery. This commit
removes the mongodb env. Setting the app itself to absent removes the
apps directory from env.d so the mongo bit causes errors when it tries
to write to a directory that no longer exists.

Once this is deployed I think we'll be able to remove this file
altogether, finally... (fingers crossed).