Disable remote syslog logging
We are using Filebeat to ship syslogs straight to our ELK stack, so we
no longer need to send to our old syslog server.

This adds a conditional and only sets it to false in AWS for now. Once
I've tested then we can remove the paramemter, set the behaviour as
default, and remove all the old machines.