Use awscli rather than s3cmd
Streaming backups for MySQL started failing for the nightly "base"
backup.

They failed because filenames were not being renamed in the S3 bucket,
but in S3 terms this is classed as a full copy and rename of an object.

This is because the `s3cmd` tool is unable to process files above 5GB. We
have only recently hit this limit for our base backups. There is an
[open issue](https://github.com/s3tools/s3cmd/issues/830) regarding this
that has not yet have a fix.

It seems that the `awscli` tool is able to handle this, although it does
also include some very verbose output which is not suitable for an
automated script.

I am not removing the `s3cmd` tool as it's used for other things within
these manifests.