Explicit link between xtrabackup interval and alert
The incremental xtrabackup check should have a freshness alert the same
as the interval at which it runs. This is currently the behaviour,
however the link between the two is not obvious. This change makes that
link more explicit.

If the alert triggers, we should investigate if it is taking too long
to run. In this case, the interval should be increased.