Remove ERRBIT_API_KEY API for all applications
This removes the ERRBIT_API_KEY variables from the configs. We don't
need it anymore because we've moved to Sentry. Note that this doesn't
remove the environment variable from the server. To do this we would
need to use `ensure => absent`. This is too much work for it to be
worth it.