Add ASSET_MANAGER_BEARER_TOKEN env var to Whitehall
In preparation for us being able to save Whitehall assets to Asset
Manager.

I've added the `asset_manager_bearer_token` info and argument such that
it appears alphabetically in the list of arguments.

I'm setting the `ASSET_MANAGER_BEARER_TOKEN` environment variable within
the `$configure_admin == true` block as I only expect whitehall-admin to
be able to write to Asset Manager.

I've tested this change in the Dev VM and confirmed that the environment
variable is written as expected.

We'll also need to ask someone to:

1. Add an application token for Asset Manager to the Whitehall API user
in signon.

2. Store the token generated in step 1 in the govuk-secrets repository.