Add service defined type
Add a new defined type for govuk_containers called "service". Similar to
the "app" defined type, this instead creates a service and should be run
on a Docker Swarm manager.

A service will not create an upstart file, and an amount of replicas can
be defined across the cluster.

Add for the release app as a test since we already have all the
environment variables set up for it.