Update filebeat options
Currently the outputs are appearing in Kibana like:

`json.@fields.foo`
`fields.application`

Ideally we do not want to have these fields prefixed as such, as per the
guidance
[here](https://gds-way.cloudapps.digital/manuals/logging.html#naming-conventions).

This sets some sensible defaults that can be overridden if required. It
sets the "keys_under_root" option for both "fields" and "json" type
logfiles.