Update monitoring URL for Logit Kibana
Create a new function that processes that unbelievably gross Kibana 5
URL syntax for a query. Seriously, there must be a different way that I
can't find documentation on.

Add a conditional that we will only enable in Integration now for
testing.

Eventually we can remove the rest of the code related to the older,
rather pleasant, syntax.