Remove direct links to Kibana 5
After fiddling around for a while, I haven't been able to find a way to
directly link to a Kibana 5 search term. The search URL ends up being so
long that Icinga truncates the link.

This is only present in a couple of alerts, so instead I will just
update the relevant documentation with instructions on how to login to
logit, and which query to run.