Upgrade Grafana via Docker
Upgrade Grafana by running it in a container rather than installing as a
package.

This is officially documented here:
http://docs.grafana.org/installation/docker/

All configuration options can be passed in as environment variables,
which makes it straight forward to pass in any additional arguments in
the future.

To upgrade we would need to update the image tag we use, which should
make life a little easier (in theory).

I considered moving all the dashboards across and removing the grafana
module, but decided it was clearer to leave them where they were. I
moved across the grafana.ini and vhost config files to be with the container.