Cut down Grafana config
Most is set as the default, and I don't think it needs to be in our
codebase.

We could potentially replace all the stuff we want to set using
environment variables and remove the need for this file completely.

Adding as a separate commit so that we can see the diff of what has been
removed.