Update query for logit use
The only index that needs changing is "@fields.application", so I did
wonder if it would be easier to add Logstash configuration that made a
duplicate field.

However, in the future we ideally want to move away from using the
`@fields` prefix (something I'm still trying to work out to do), so I
think it's best to put in a conditional until we've entirely migrated to
Logit.

I've tested the datasource and it correctly works, but it does require
Logit specifically enabling it for us via a support ticket (The Future
TM).