Add a temporary directory for Whitehall / Asset Manager
We are adding the ability for Whitehall to upload non-attachment
assets via the Asset Manager API. For this to work, the requests to
the Asset Manager API need to be made in Sidekiq workers rather than
in-request. This means we need some shared state (i.e the
`/mnt/uploads` NFS mount) between the machines to store the files
temporarily so that the upload and worker processes are able to run on
different machines.