Remove outer add_header for X-Frame-Options in AssetManager Nginx config
I should've done this as part of #6602. Since that PR, the two location
blocks which are used to serve the only public URLs from the Asset
Manager app, both set the X-Frame-Options response header explicitly, so
the "outer" `add_header` statement is adding a spurious extra value for
this header.