Remove logstream shipping logs
Use logstream to only ship metrics, so unless statsd_timer or
statsd_metric are defined, set the file to absent.

Remove the parameter regarding `json => true` to ensure Puppet does not
complain. We'll leave the resources in place though to ensure that the
files are appropriately removed, and then remove the code later.