Ensure Kibana app only redirects to Logit docs
We will need to remove the rest of the code when it has been deployed to
Production. The only thing the kibana manifest should do is redirect
people to our documentation regarding Logit.

In the future we're hoping Logit can provide unique links to stacks, so
we'll be able to set up a better redirect.

We need to do this in two stages: first, ensure that the resources are
set to absent to remove code. Then when deployed, remove the rest of the
code and add the redirect back in. This is due to a duplicate resource
name.