Add hack for removing logstream service
This errored because `initctl stop` errored, which meant the dependency
to remove the file never runs.

If the file does not exist, do not run the file or exec resources, but
if it does, ensure the service is stopped but do not error if it already
is stopped, and then remove the file.