Remove deployment dependency on github
This changes the deploy app job to no longer have a SCM association with
github.com:alphagov/govuk-app-deployment. Instead the repository that is
associated with this is determined based on whether the deploy with
gitlab checkbox is selected or not.

We specify the banch of master even though it is not required so that it
will be clear for future users of this script where to change this when
you want to use different branches of govuk-app-deployment.