Update swarm logic
My original logic assumed a single manager. If the manager dies, then
all state is lost and the swarm must be recreated and all apps would
need to be redeployed.

As such, we should have at least 3 managers. This allows the logic for a
new manager to join a cluster if it dies and is recreated.