Remove apps that are no longer deployable
With the retiring of [alphagov-deployment][] these apps will no longer
be safe to deploy as they won't have files automatically copied into
them. Therefore to avoid accidental incorrect deployments they are
removed from the list of deployable applications.

If someone was in a bind it'd still be easy to deploy an app as you'd
just have to edit the jenkins job to add the application to the list.

[alphagov-deployment]: https://github.com/alphagov/alphagov-deployment