Remove legacy logging
We are doing work to ensure all applications log in a consistent manner.

Currently, depending on how they're configured, they log to either:

`/var/app/$application/log/production.log`
`/var/app/$application/log/production.json.log`
`/var/log/$application/app.out.log`

We had a conditional that we would have to change when an app has been
migrated over. We're looking to centralise configuration and update
apps, and so it would be easier to just try logging from both places. It
should not cause any errors if the log files are empty, it just won't
log anything.

This will make it easier to switch over and ensure we're not losing any
logs when we migrate.

When this has been completed, we should remove all code related to
`production.log` and `production.json.log`, and ensure that the only places
that apps log are `/var/log/$application`.