Add streaming RDS backup scripts
Add backup scripts that do not do anything too clever: take a list of
databases, excluding some specific ones, take a dump, and stream the
output to a file in an S3 bucket. All backups will go into the same
bucket, which we'll enable with expiration dates to manage storage.

Ensure there are passive checks that alert us if a backup fails.

In theory these scripts could also be run on demand if developers wanted
more up to date data to import.

We are relying on the connection during transfer being encrypted (which
is by default with gof3r), and the bucket being encrypted at rest.

We will use IAM roles to allow access to the bucket from the instance.