Only configure Sendfile for public asset requests from Asset Manager
Previously we were setting the X-Sendfile-Type & X-Accel-Mapping request
headers in *all* requests proxied from Nginx to the Rails app. However,
these headers only need to be set for the /media/* (Mainstream) and
/government/uploads/* (Whitehall) URL paths, because these are the only
routes which ever make use of the Sendfile behaviour.

In the not too distant future, we're planning to change the Rails app to
proxy all these public asset requests to S3, like public Mainstream
asset requests already are in production. At this point, it will be
possible to remove both this new Nginx location block and the related
internal location block.

My iniital motivation for doing this was to solve a problem with
alphagov/asset-manager#267 in which the Rails app proxies public asset
requests to fake S3 URLs served by the Rails app. Having the
Sendfile-related headers set for the latter requests was problematic
for serving the fake S3 URLs.

Note that I have introduce a bit of duplication in the new Nginx
location block (in the proxy_pass statement). I did try referencing the
existing named location block "app", but I couldn't get this to work.
However, given that the amount of duplication is minimal, there is
already duplication like this in the Nginx config, and the location
block should be going away in the not too distant future, I'm happy to
leave it like this.