Fail fast if Nginx config contains a syntax error
Previously if the Nginx config changed (e.g. in an individual virtual
host config), but it contained a syntax error, the following would occur
when running `govuk_puppet`:

1. Puppet notices the config has changed and triggers a restart of the
Nginx service
2. Puppet runs `/etc/init.d/nginx reload` to restart the Nginx service
3. The syntax error prevents the new config from being loaded
4. Nginx continues to run with the old config
5. `/etc/init.d/nginx reload` returns a zero (success) error code
6. Puppet assumes the restart has completed successfully and continues
without any warnings or errors
7. If the Nginx process is subsequently restarted (e.g. on reboot), it
will stop and then fail to start, because of the syntax error

After this change, Puppet's restart of the Nginx service first checks
that there are no syntax errors in the config and fails fast with a
useful error message if there is a syntax error. Otherwise it continues
as before.

Ideally we'd also roll back the change to the Nginx configuration, but
I'm not sure how to do that. However, I still think that this change is
an improvement, because (assuming we have alerts setup for the Puppet
cron job failing), we'll find out immediately if there is a syntax error
in the Nginx config.

The motivation behind this change was a real case where a syntax error
in the Nginx config was accidentally introduced and Nginx silently
failed to reload the config. Later when the (integration) machine was
rebooted, the OS attempted to start Nginx, but it failed to start,
meaning that all requests to any Nginx virtual hosts failed.

Note that this is similar to the config check that happens in the
`nginx::restart` class, although as @chrisroos has pointed out [1], I
don't think is is working as intended either.

[1]:
https://github.com/alphagov/govuk-puppet/pull/6693#issuecomment-339245051