Create mongo backups to S3
We already have scripts and manifests that create backups for MongoDB to
an S3 bucket. However, these scripts are heavily based on a few
assumptions:

 - We want to encrypt using GPG
 - We are not in an AWS environment
 - Our hosts are static

Rather than try to work around all the code that does the above, I
created a new class and template which excludes anything to do with GPG
and AWS API settings.

Rather than also having to set a load of hieradata to specifically
include the class of a specific node, I just include it on all nodes,
but only run it if the instance is the primary instance.

We are unable to specify based upon hostname in AWS, and this also
ensures that if primary goes away in our cluster, when the new primary
steps up, backups continue to happen.

This means that a Nagios passive check is included on all machines, but
they should always report back as "OK" if they're not a primary so
should not cause any alarms.