Add number of Sentry errors to the 2nd line dashboard
This adds a metric with the number of errors reported in Sentry over
the last hour. It gets this from
`stats.gauges.sentry-error-count-last-hour.(appname)`, which is
populated every 10 minutes by this Jenkins job:

https://deploy.publishing.service.gov.uk/job/govuk-sentry-monitor

Which runs the script in this repo:

https://github.com/alphagov/govuk-sentry-monitor

The intention is that we'll some visibility on the error reporting. We
started paying per-error since the move to Sentry so preventing some
apps from eating up all our budget is important.

https://trello.com/c/btVExJP7