Remove elasticsearch plugin
When run on integration, this gives " ERROR: failed to download out of
all possible locations..., use --verbose to get detailed information"
and changing the `module_dir` to match how we had it previously
didn't make a difference.

I'm not sure what the puppet provider is doing, as it works fine
from the CLI with the same URL.

I'm just going to remove this for now as it's not important right now and
I can install through the CLI on integration anyway.