Make the integration_router_data_deploy job actually do something
When this was added in 3fa8e443ed7a6b1117adcf0429c25da338967596 the job
only `echo`ed the curl to deploy.integration rather than actually calling
it.  This means the job is a no-op and doesn't actually trigger the
deploy.  We remove the echo so it does trigger the deploy and while we're
at it we also add the CRUMB stuff added to the integration_deploy_app job
in #6585.