Allow supplying the repo name to buildProject
In some cases the repository name for a project and the name of the
jenkins job that runs the build are different.  For example the
Contacts Admin publishing app lives in a repo called contacts-admin, but
the jenkins job is called contacts (as are many other puppet references
to this app).  This means that the contacts-admin app can't use the
standard buildProject method in its Jenkinsfile because this assumes that
the repo matches the job.

It's not 100% clear when we should use the repoName and when we should
use the jobName.  For example the publishGem and pushDockerImage methods
take a name, but it's not actually important that it's the repository or
jenkins job name because in neither case does it make a reference to
github or jenkins.  Where it's not obvious I've left the param as the
jobName to preserve existing behaviour.