Add pseudo-app to setup env for syncing assets from production
The new `asset_env_sync` class declares the dependencies which the
`env-sync-and-backup` project will need in order to sync assets from the
production S3 bucket to the staging & integration S3 buckets. That
project will use the AWS command line tool to do the sync operation. In
order to be able to do this, it will need the credentials for the
`asset-manager-production-env-sync-user` IAM user which has read access
to the production S3 bucket and write access to the staging &
integration S3 bcukets. These credentials are made available in the
`asset-env-sync` environment in the usual way.

The new pseudo-app, `asset_env_sync` is allocated to the
`asset-master-1` node in all environments (except development), but is
only enabled in production, because the latter is the only environment
in which the `env-sync-and-backup` project needs to execute the sync
operation.