AWS remove old sshkey resources
On AWS machines are created and destroyed quite often, generating
conflicting entries in the known_hosts file when IPs are reused.

The `sshkey` resource that we export and collect doesn't remove
old entries automatically from the known_hosts file, even when resources
are no longer present in PuppetDb. This change forces the resource
to be purged.