Enable Asset Manager proxying to S3 in all environments
This was enabled in production in these PRs [1,2] and has been working
successfully for some time. However, we haven't been able to enable it
in staging & integration, because we haven't had the nightly syncs of
Asset Manager assets from the production S3 bucket to the integration &
staging buckets.

Since some Whitehall assets are now served via Asset Manager and their
URLs are environment-specific, not having the nightly syncs in place
would've meant existing Whitehall assets wouldn't have been served
successfully via Asset Manager in integration & staging.

We're planning to enable the nightly syncs shortly in the PRs mentioned
in this comment [3]. At that point, we should be able to merge & deploy
the changes in this commit.

We haven't been able to enable it in development, because we felt it was
too onerous to require developers to setup the relevant S3 resources
just for a development environment. However, in this PR [4] we
implemented a "fake" S3 within the Rails app for just this purpose.

[1]: https://github.com/alphagov/govuk-puppet/pull/6432
[2]: https://github.com/alphagov/govuk-puppet/pull/6736
[3]:
https://github.com/alphagov/asset-manager/issues/145#issuecomment-342894788
[4]: https://github.com/alphagov/asset-manager/pull/267