Allow providing extra parameters to buildProject
This allows apps that use custom jenkins build parameters to use the
standard buildProject script for their builds.  For example smart-answers
has a `RUN_REGRESSION_TESTS` parameter that tells the build to run an
extra set of tests (which are very slow, so not run by default) as part of
the build.  The `properties` call that we use to define these parameters
can't be called multiple times to define the params so we need to add that
extra parameter to the `properties` call in here or smart-answers can't
use it.