Enable Gor for AWS on cache-1 staging
Start replaying traffic on AWS Integration from cache-2 on Staging.
Add a new parameter to router::gor to disable adding the target IP
to the hosts file for the AWS endpoint (it's behind an ELB so we need
to resolve). On Carrenza we keep the current configuration, adding the
IP to the hosts file, to avoid making DNS calls.