Update hostalias and display_names of hosts in AWS
The idea behind this commit is being able to sort hosts in the Icinga
dashboard in alphabetical order.

Currently the display name gives us a friendly name, but because the
underlying host is something like "ip-10-1-6-6.eu-west-1.compute.internal" the
hosts are not sorted in a friendly manner. Sorting hosts in a friendly
way makes it easy to see if all hosts of the same type are suffering
from the same issue, or if it is only a single host that has a problem.

To ensure that the display_name matches the underlying host, the
"hostalias" needs to be updated. This updates the full host naming
within Icinga, so hosts will be reported as:

`graphite_ip-10-1-6-6`