Fix mongodump-to-s3 backups again
It appears my bugfix in 524e056d51a348667482fbccc7ca34a929b811a1 was
incorrect.

I was caught out by differences between the awscli tool on my work
client, and the version on the server.

The error was not related to where the argument was placed, but rather
providing the encryption after the argument. Not setting the type of
encryption defaults to AES256, and I have successfully tested and pushed
files to an S3 bucket with this command.