Update Grafana dashboards for AWS
Our Grafana dashboards are using hostname regular expressions that
don't work on AWS. On Carrenza the metric fqdn is created including the
VPC name, which is not part of the AWS infrastructure (on AWS we use
`stackname_environment` as part of the metric fqdn), so the VPC part
needs to be removed from the expression.

All the `*_frontend` expressions need to be updated to explicity exclude efg
and licensing (if we use `*frontend*` we would be including efg and
licensing on the Carrenza dashboards, which is not bad but currently
are not included).