Revert "Merge pull request #6769 from alphagov/enable-asset-manager-proxying-to-s3-via-nginx-in-integration-and-staging"
It turns out that assets sync-ed from the production S3 bucket to the
staging & production buckets have a different owner and therefore are
not accessible by the Asset Manager.

This only affects Whitehall assets served from Asset Manager which is
currently only the Organisation Logos. This is because the URLs for
Mainstream assets are embedded in content in the Content Store and the
data for the latter is sync-ed from production to staging & integration.
This in turn means that URLs for existing Mainstream assets are always
production URLs and those are working OK, because objects are not
sync-ed into the production S3 bucket.

The problem also only affects existing assets; not new assets uploaded
in the relevant environment.

Although the problem is small, it seems sensible to revert these changes
while we come up with a fix.

This reverts commit 47797a44cc8683ea742c7653bc2a3786ba9e6a0a, reversing
changes made to 87c24a46b512e12991fd9960032bf297ef26b84c.