Create jobs that deploy all apps for a given node
This is to allow easy deployment of all apps for a given node class. The
context behind this is that if a node goes away in AWS, it will be
rebuilt by the autoscaling group. If this occurs, then the applications
will not be automatically deployed, and having to log in to deploy them
all is problematic to do by hand.

This is a very short term solution that creates jobs for each type of
node class, that triggers downstream jobs for each app that is on that
node class. It will use the "deployed-to-integration" branch which is
set whenever an app is deployed to determine what version of the
application is running in each environment.

It will also deploy using the task "deploy:without_migrations", as we
just want to deploy the code automatically rather than running any tasks
against the database.

This does also kick off the Smokey job after each successful deployment
of an app, and this takes ~6 minutes to run, but I don't think this is
too problematic.

There is a blacklist of apps that are not "deployable" through Jenkins,
and having these created as upstream jobs will cause the job itself to
fail, so they must be removed.