Fix worker panels
The ref_id's need to be unique otherwise the later one overwrites the former.
This results in the worker panels on dashboards not working for example on the
Publisher dashboard.

It appears that you can leave them out altogether, but this is not the
convention used by other configuration, so we'll just use A and B.