Change search index checks to use absolute thresholds
Using relative values is confusing and error prone, so just do a simple diff.

The check always compares to 3 days ago. This is so that on monday morning, it
will compare to friday, instead of the weekend. If the alert triggers out of
hours, we don't want that value to become the new baseline until we've had a
chance to see it.

If statsd data is missing (a value of None), no alert will be raised.
This is not a huge problem. As long as statsd is running it keeps gauges at
their previous value:
https://github.com/etsy/statsd/blob/master/docs/metric_types.md#gauges