Add a separate jenkins job for recreating the search indexes
This process
- uses rummager's configuration to generate the settings/mappings for a
  new set of search indexes
- creates the new indexes
- triggers a reindex of content from the old index to the new one
- while this is happening, a write lock is placed on the old index so
  data isn't lost
- after completing the reindex, run some checks on the new index, and
  switch over to it if successful (by moving aliases)

This used to be part of the search_fetch_analytics job, and ran every
night but we split it out, and now only need to run it when changing
the mappings or updating elasticsearch.

This is a very slow process, so we don't want to hook it into our CI
pipeline, but we do want to test it periodically on integration.