Add docker as an apt-mirror
With the release of docker-ce docker is now it's own repository.
See: https://docs.docker.com/engine/installation/linux/docker-ce/ubuntu/#set-up-the-repository
for more details

I've added repos of stable as there are directories of different
releases inside
https://download.docker.com/linux/ubuntu/dists/trusty/

and the key is based on the one provided on the above link.