Add environment filter to sentry link on deployment dashboard
Otherwise we link to all errors for the project.  Sentry doesn't make it
obvious via the UI how to filter to only see the errors for a given
environment so we might mean people don't check for errors because it's
too noisy, or worry about errors on integration when they're deploying
to production.

We use the `govuk::deploy::config::errbit_environment_name` hiera as the
default value for the sentry environment because that's currently the name
of the puppet var used to set the environment variable used to configure
sentry in our apps.