Avoid S3 HTTP headers appearing in Asset Manager responses
* `x-amz-replication-status` started appearing when we enabled
cross-region replication on the production S3 bucket.

* `x-amz-meta-md5-hexdigest` has been appearing since we added this
piece of custom metadata in this commit [1].

This commit hides the above response headers when Asset Manager proxies
asset requests to S3 via Nginx which it is now doing by default in all
environments.

This addresses both this issue [2] and this other issue [3].

[1]:
https://github.com/alphagov/asset-manager/commit/ff2ed6163e80327d5b51f0cabfe16b8013601a84
[2]: https://github.com/alphagov/asset-manager/issues/270
[3]: https://github.com/alphagov/asset-manager/issues/289